ugly fix

+++ ManagementSystem/src/app/routes/dashboard/system-parameter/parameter-list/parameter-list.component.ts
@@ -1,11 +1,17 @@
-import { HttpClient } from '@angular/common/http';
-import { Component, OnInit, ViewChild, ViewEncapsulation } from '@angular/core';
-import { NbToastrService } from '@nebular/theme';
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
+import { Component, OnDestroy, OnInit, ViewChild, ViewEncapsulation } from '@angular/core';
+import { NbToastrService, NbWindowService } from '@nebular/theme';
 import { ServerDataSource } from 'ng2-smart-table';
+import { interval, Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 import { SystemParameter } from 'src/app/entity';
 import { RESTfulAPI } from 'src/app/service/restful';
 import { SysParamService } from 'src/app/service/system-param/system-parameter.service';
-import { Pattern } from 'src/app/shared/utils';
+import { ConfirmWindowComponent } from 'src/app/shared/components/confirm-window.component';
+import { Pattern, UglyInputHint } from 'src/app/shared/utils';
+
+
+function isEmpty(v: string | null) {return !v || v == '';}
 
 @Component({
     selector: 'ngx-parameter-list',
@@ -13,7 +19,8 @@ import { Pattern } from 'src/app/shared/utils';
     styleUrls: ['./parameter-list.component.scss'],
     encapsulation: ViewEncapsulation.None,
 })
-export class ParameterListComponent implements OnInit {
+export class ParameterListComponent implements OnInit, OnDestroy {
+    destroy$: Subject<void>;
 
     // settings //{
     settings = {
@@ -47,24 +54,22 @@ export class ParameterListComponent implements OnInit {
             confirmDelete: true,
         },
         columns: {
-            key: {
+            remark: {
                 title: '名称',
                 filter: false,
                 type: 'text',
+            },
+            key: {
+                title: '关键字',
+                filter: false,
+                type: 'text',
                 editable: false,
+                class: 'fucking-youyou',
             },
             value: {
                 title: '值',
                 filter: false,
                 type: 'text',
-            },
-            remark: {
-                title: '备注',
-                filter: false,
-                type: 'text',
-                editor: {
-                    type: 'textarea'
-                },
             },
         },
     }; //}
@@ -72,7 +77,13 @@ export class ParameterListComponent implements OnInit {
     source: ServerDataSource;
     constructor(private toastrService: NbToastrService,
                 private sysParamService: SysParamService,
-                private http: HttpClient) {}
+                private http: HttpClient,
+                private windowService: NbWindowService) {}
+
+    ngOnDestroy(): void {
+        this.destroy$.next();
+        this.destroy$.complete();
+    }
 
     ngOnInit(): void {
         this.source = new ServerDataSource(this.http, {
@@ -86,14 +97,41 @@ export class ParameterListComponent implements OnInit {
             totalKey: 'total',
         });
 
-        (this.source as any).find = function (element: SystemParameter): Promise<any> {
-            const found = (this.data as SystemParameter[]).find(elem => elem.key == element.key);
+        (this.source as any).find = function (element: SystemParameter | string): Promise<any> {
+            if (typeof element == 'object') {
+                element = element.key;
+            }
+            const found = (this.data as SystemParameter[]).find(elem => elem.key == element);
             if(found) {
                 return Promise.resolve(found);
             } else {
                 return Promise.reject("NOT FOUND");
             }
         }
+
+        this.destroy$ = new Subject();
+        interval(1000).pipe(takeUntil(this.destroy$))
+            .subscribe(() => {
+                UglyInputHint(input => input.placeholder == '关键字',
+                              async input => {
+                                  if (isEmpty(input.value)) {
+                                      return '此字段不为空';
+                                  }
+
+                                  if (!this.keyRegex.test(input.value)) {
+                                      return this.keyRegex[Pattern.HintSym];;
+                                  }
+
+                                  try {
+                                      if(await this.source.find(input.value) != null)
+                                          return '系统参数已存在';
+                                  } catch {}
+
+                                  return null;
+                              },
+                              input => isEmpty(input.value) && '此字段不为空' || null,
+                );
+            });
     }
 
     onsearchinput(pair: [string, (hints: string[]) => void]) //{
@@ -107,6 +145,9 @@ export class ParameterListComponent implements OnInit {
             data.forEach((value) => {
                 if(value.key.match(input)) {
                     ans.push(value.key);
+                }
+                if(value.remark.match(input)) {
+                    ans.push(value.remark);
                 }
             })
         }
@@ -123,17 +164,24 @@ export class ParameterListComponent implements OnInit {
         this.source.refresh();
     } //}
 
+    private nameRegex = Pattern.Regex.uname;
     private keyRegex = Pattern.Regex.aname;
     private async checkData(row: SystemParameter, toastr: boolean = true): Promise<boolean> //{
     {
-        if(this.keyRegex.test(row.key)) {
-            return true;
-        } else {
-            if(toastr) {
-                this.toastrService.warning(`非法名称字段: ${this.keyRegex[Pattern.HintSym]}`, "系统参数");
-            }
+        if(isEmpty(row.remark) || isEmpty(row.key) || isEmpty(row.value)) {
+            this.toastrService.warning(`名称、关键字和值必填`, "系统参数");
             return false;
         }
+
+        if(!this.nameRegex.test(row.remark) && toastr) {
+            this.toastrService.warning(`非法名称字段: ${this.nameRegex[Pattern.HintSym]}`, "系统参数");
+            return false;
+        }
+        if(!this.keyRegex.test(row.key)) {
+            this.toastrService.warning(`非法关键字字段: ${this.keyRegex[Pattern.HintSym]}`, "系统参数");
+            return false;
+        }
+        return true
     } //}
 
 
@@ -151,8 +199,13 @@ export class ParameterListComponent implements OnInit {
         try {
             await this.sysParamService.post(event.newData);
             this.toastrService.success(`新建'${event.newData.key}'`, "系统参数");
-        } catch {
-            this.toastrService.danger(`创建系统参数'${event.newData.key}'失败`, "系统参数");
+        } catch (e: any) {
+            let errorMsg = `创建系统参数'${event.newData.key}'失败`;
+            if (e instanceof HttpErrorResponse) {
+                errorMsg = e.error.reason || errorMsg;
+
+            }
+            this.toastrService.danger(errorMsg, "系统参数");
             return event.confirm.reject();
         }
 
@@ -175,8 +228,12 @@ export class ParameterListComponent implements OnInit {
         try {
             await this.sysParamService.put(event.newData);
             this.toastrService.success(`修改'${event.data.key}'`, "系统参数");
-        } catch {
-            this.toastrService.danger(`编辑系统参数'${event.newData.key}'失败`, "系统参数");
+        } catch (e: any) {
+            let errorMsg = `编辑系统参数'${event.newData.key}'失败`;
+            if(e instanceof HttpErrorResponse) {
+                errorMsg = e.error.reason || errorMsg;
+            }
+            this.toastrService.danger(errorMsg, "系统参数");
             return event.confirm.reject();
         }
 
@@ -191,10 +248,22 @@ export class ParameterListComponent implements OnInit {
     }) //{
     {
         try {
+            const win = this.windowService.open(ConfirmWindowComponent, {
+                title: `删除系统参数`,
+                context: {message: `确认删除关键字为 '${event.data.key}' 的系统参数?`}
+            });
+            await win.onClose.toPromise();
+            if (!win.config.context['isConfirmed'])
+                return;
+
             await this.sysParamService.delete(event.data.key);
             this.toastrService.success(`删除'${event.data.key}'`, "系统参数");
-        } catch {
-            this.toastrService.danger(`删除系统参数'${event.data.key}'失败`, "系统参数");
+        } catch (e: any) {
+            let errorMsg = `删除系统参数'${event.data.key}'失败`;
+            if (e instanceof HttpErrorResponse) {
+                errorMsg = e.error.reason || errorMsg;
+            }
+            this.toastrService.danger(errorMsg, "系统参数");
             return event.confirm.reject();
         }
